+++ R/pdf_create_blank.R
@@ -37,7 +37,7 @@ pdf_create_blank <- function(
 		on.exit(dev.set(current_dev), add = TRUE)
 	}
 
-	if (missing(width) && missing(height)) {
+	if (missing(width) && missing(height) && paper != "special") {
 		pnp_pdf(
 			output,
 			paper = paper,