+++ NEWS.md
@@ -6,7 +6,7 @@ Breaking changes
 
 * The first argument of `layout_preset()` is now `preset` instead of `name` (so we can now pass `layout_grid()`'s new `name` argument in the `...`).
 * The default argument for `paper` is now `getOption("papersize", "letter")` instead of (depending on the function) "letter" or "special".
-* `bm_create_pdf()` and `pdf_create_blank()` now follow the convention that paper size will be determined by `height` and `width` if either are non-missing otherwise solely by `paper` and `orientation`.
+* `bm_create_pdf()` and `pdf_create_blank()` now follow the convention that paper size will be determined by `height` and `width` if either are non-missing (and/or `paper = "special"`) otherwise paper size will be determined by both `paper` and `orientation`.
 
 New features
 ------------